+++ src/main.cc
@@ -17,6 +17,7 @@
  */
 
 #include "class_diagram/generators/plantuml/class_diagram_generator.h"
+#include "cli/cli_handler.h"
 #include "config/config.h"
 #include "include_diagram/generators/plantuml/include_diagram_generator.h"
 #include "package_diagram/generators/plantuml/package_diagram_generator.h"
@@ -30,8 +31,6 @@
 #include <backward-cpp/backward.hpp>
 #endif
 
-#include <clang/Basic/Version.h>
-#include <clang/Config/config.h>
 #include <cli11/CLI11.hpp>
 #include <spdlog/spdlog.h>
 
@@ -49,51 +48,6 @@ backward::SignalHandling sh; // NOLINT
 #endif
 
 using namespace clanguml;
-
-/**
- * Print the program version and basic information
- */
-void print_version();
-
-/**
- * Print list of diagrams available in the configuration file
- *
- *  @param cfg Configuration instance loaded from configuration file
- */
-void print_diagrams_list(const clanguml::config::config &cfg);
-
-/**
- * Print effective config after loading and setting default values.
- *
- *  @param cfg Configuration instance loaded from configuration file
- */
-void print_config(const clanguml::config::config &cfg);
-
-/**
- * Generate sample configuration file and exit.
- *
- * @return 0 on success or error code
- */
-int create_config_file();
-
-/**
- * Add example diagram of given type to the config file.
- *
- * @param type Type of the sample diagram to add
- * @param config_file_path Path to the config file
- * @param name Name of the new diagram
- * @return 0 on success or error code
- */
-int add_config_diagram(clanguml::common::model::diagram_t type,
-    const std::string &config_file_path, const std::string &name);
-
-/**
- * Check if diagram output directory exists, if not create it
- *
- * @param dir Path to the output directory
- * @return True if directory exists or has been created
- */
-bool ensure_output_directory_exists(const std::string &dir);
 
 /**
  * Generate specific diagram identified by name
@@ -156,145 +110,22 @@ void generate_diagrams(const std::vector<std::string> &diagram_names,
 
 int main(int argc, const char *argv[])
 {
-    CLI::App app{"Clang-based PlantUML diagram generator for C++"};
+    cli::cli_handler cli;
+    auto res = cli.handle_options(argc, argv);
 
-    std::string config_path{".clang-uml"};
-    std::optional<std::string> compilation_database_dir{};
-    std::vector<std::string> diagram_names{};
-    std::optional<std::string> output_directory{};
-    unsigned int thread_count{0};
-    bool show_version{false};
-    int verbose{0};
-    bool list_diagrams{false};
-    bool quiet{false};
-    bool initialize{false};
-    std::optional<std::string> add_class_diagram;
-    std::optional<std::string> add_sequence_diagram;
-    std::optional<std::string> add_package_diagram;
-    std::optional<std::string> add_include_diagram;
-    bool dump_config{false};
-    std::optional<bool> paths_relative_to_pwd{};
-
-    app.add_option(
-        "-c,--config", config_path, "Location of configuration file");
-    app.add_option("-d,--compile-database", compilation_database_dir,
-        "Location of compilation database directory");
-    app.add_option("-n,--diagram-name", diagram_names,
-        "List of diagram names to generate");
-    app.add_option("-o,--output-directory", output_directory,
-        "Override output directory specified in config file");
-    app.add_option("-t,--thread-count", thread_count,
-        "Thread pool size (0 = hardware concurrency)");
-    app.add_flag("-V,--version", show_version, "Print version and exit");
-    app.add_flag("-v,--verbose", verbose,
-        "Verbose logging (use multiple times to increase - e.g. -vvv)");
-    app.add_flag("-q,--quiet", quiet, "Minimal logging");
-    app.add_flag("-l,--list-diagrams", list_diagrams,
-        "Print list of diagrams defined in the config file");
-    app.add_flag("--init", initialize, "Initialize example config file");
-    app.add_option(
-        "--add-class-diagram", add_class_diagram, "Add class diagram config");
-    app.add_option("--add-sequence-diagram", add_sequence_diagram,
-        "Add sequence diagram config");
-    app.add_option("--add-package-diagram", add_package_diagram,
-        "Add package diagram config");
-    app.add_option("--add-include-diagram", add_include_diagram,
-        "Add include diagram config");
-    app.add_flag(
-        "--dump-config", dump_config, "Print effective config to stdout");
-    app.add_flag("--paths-relative-to-pwd", paths_relative_to_pwd,
-        "If true, all paths in configuration files are relative to the $PWD "
-        "instead of actual location of `.clang-uml` file.");
-
-    CLI11_PARSE(app, argc, argv);
-
-    if (show_version) {
-        print_version();
+    if (res == cli::cli_flow_t::kExit)
         return 0;
-    }
 
-    if (initialize) {
-        return create_config_file();
-    }
-
-    verbose++;
-
-    if (quiet)
-        verbose = 0;
-
-    if (add_class_diagram) {
-        return add_config_diagram(clanguml::common::model::diagram_t::kClass,
-            config_path, *add_class_diagram);
-    }
-
-    if (add_sequence_diagram) {
-        return add_config_diagram(clanguml::common::model::diagram_t::kSequence,
-            config_path, *add_sequence_diagram);
-    }
-
-    if (add_package_diagram) {
-        return add_config_diagram(clanguml::common::model::diagram_t::kPackage,
-            config_path, *add_package_diagram);
-    }
-
-    if (add_include_diagram) {
-        return add_config_diagram(clanguml::common::model::diagram_t::kInclude,
-            config_path, *add_include_diagram);
-    }
-
-    clanguml::util::setup_logging(verbose);
-
-    clanguml::config::config config;
-    try {
-        config = clanguml::config::load(config_path, paths_relative_to_pwd);
-    }
-    catch (std::runtime_error &e) {
-        LOG_ERROR(e.what());
-        return 1;
-    }
-
-    if (list_diagrams) {
-        print_diagrams_list(config);
-        return 0;
-    }
-
-    LOG_INFO("Loaded clang-uml config from {}", config_path);
-
-    //
-    // Override selected config options from command line
-    //
-    if (compilation_database_dir) {
-        config.compilation_database_dir.set(
-            util::ensure_path_is_absolute(compilation_database_dir.value())
-                .string());
-    }
-    if (output_directory) {
-        config.output_directory.set(
-            util::ensure_path_is_absolute(output_directory.value()).string());
-    }
-
-    LOG_INFO("Loading compilation database from {} directory",
-        config.compilation_database_dir());
-
-    auto od = config.output_directory();
-    if (output_directory)
-        od = output_directory.value();
-
-    if (dump_config) {
-        print_config(config);
-        return 0;
-    }
-
-    if (!ensure_output_directory_exists(od))
+    if (res == cli::cli_flow_t::kError)
         return 1;
 
     std::string err{};
     auto db = clang::tooling::CompilationDatabase::autoDetectFromDirectory(
-        config.compilation_database_dir(), err);
+        cli.config.compilation_database_dir(), err);
 
     if (!err.empty()) {
         LOG_ERROR("Failed to load compilation database from {}",
-            config.compilation_database_dir());
+            cli.config.compilation_database_dir());
         return 1;
     }
 
@@ -307,10 +138,11 @@ int main(int argc, const char *argv[])
     // We have to generate the translation units list for each diagram before
     // scheduling tasks, because std::filesystem::current_path cannot be trusted
     // with multiple threads
-    find_translation_units_for_diagrams(diagram_names, config,
+    find_translation_units_for_diagrams(cli.diagram_names, cli.config,
         compilation_database_files, translation_units_map);
 
-    generate_diagrams(diagram_names, config, od, db, verbose, thread_count,
+    generate_diagrams(cli.diagram_names, cli.config,
+        cli.effective_output_directory, db, cli.verbose, cli.thread_count,
         translation_units_map);
 
     return 0;
@@ -475,189 +307,4 @@ void find_translation_units_for_diagrams(
             translation_units_map[name] = std::move(valid_translation_units);
         }
     }
-}
-
-bool ensure_output_directory_exists(const std::string &dir)
-{
-    namespace fs = std::filesystem;
-    using std::cout;
-
-    fs::path output_dir{dir};
-
-    if (fs::exists(output_dir) && !fs::is_directory(output_dir)) {
-        cout << "ERROR: " << dir << " is not a directory...\n";
-        return false;
-    }
-
-    if (!fs::exists(output_dir)) {
-        return fs::create_directories(output_dir);
-    }
-
-    return true;
-}
-
-void print_version()
-{
-    constexpr auto kLLVMBackendPackageStringLength{5};
-    std::cout << "clang-uml " << clanguml::version::CLANG_UML_VERSION << '\n';
-    std::cout << "Copyright (C) 2021-2023 Bartek Kryza <bkryza@gmail.com>"
-              << '\n';
-    std::cout << "Built against LLVM/Clang libraries version: "
-              << std::string{BACKEND_PACKAGE_STRING}.substr(
-                     kLLVMBackendPackageStringLength)
-              << std::endl;
-    std::cout << "Using LLVM/Clang libraries version: "
-              << clang::getClangFullVersion() << std::endl;
-}
-
-void print_diagrams_list(const clanguml::config::config &cfg)
-{
-    using std::cout;
-
-    cout << "The following diagrams are defined in the config file:\n";
-    for (const auto &[name, diagram] : cfg.diagrams) {
-        cout << "  - " << name << " [" << to_string(diagram->type()) << "]";
-        cout << '\n';
-    }
-}
-
-int create_config_file()
-{
-    namespace fs = std::filesystem;
-    using std::cout;
-
-    fs::path config_file{"./.clang-uml"};
-
-    if (fs::exists(config_file)) {
-        cout << "ERROR: .clang-uml file already exists\n";
-        return 1;
-    }
-
-    YAML::Emitter out;
-    out.SetIndent(2);
-    out << YAML::BeginMap;
-    out << YAML::Comment("Change to directory where compile_commands.json is");
-    out << YAML::Key << "compilation_database_dir" << YAML::Value << ".";
-    out << YAML::Newline
-        << YAML::Comment("Change to directory where diagram should be written");
-    out << YAML::Key << "output_directory" << YAML::Value << "docs/diagrams";
-    out << YAML::Key << "diagrams" << YAML::Value;
-    out << YAML::BeginMap;
-    out << YAML::Key << "example_class_diagram" << YAML::Value;
-    out << YAML::BeginMap;
-    out << YAML::Key << "type" << YAML::Value << "class";
-    out << YAML::Key << "glob" << YAML::Value;
-    out << YAML::BeginSeq << "src/*.cpp" << YAML::EndSeq;
-    out << YAML::Key << "using_namespace" << YAML::Value;
-    out << YAML::BeginSeq << "myproject" << YAML::EndSeq;
-    out << YAML::Key << "include";
-    out << YAML::BeginMap;
-    out << YAML::Key << "namespaces";
-    out << YAML::BeginSeq << "myproject" << YAML::EndSeq;
-    out << YAML::EndMap;
-    out << YAML::Key << "exclude";
-    out << YAML::BeginMap;
-    out << YAML::Key << "namespaces";
-    out << YAML::BeginSeq << "myproject::detail" << YAML::EndSeq;
-    out << YAML::EndMap;
-    out << YAML::EndMap;
-    out << YAML::EndMap;
-    out << YAML::EndMap;
-    out << YAML::Newline;
-
-    std::ofstream ofs(config_file);
-    ofs << out.c_str();
-    ofs.close();
-
-    return 0;
-}
-
-int add_config_diagram(clanguml::common::model::diagram_t type,
-    const std::string &config_file_path, const std::string &name)
-{
-    namespace fs = std::filesystem;
-
-    fs::path config_file{config_file_path};
-
-    if (!fs::exists(config_file)) {
-        std::cerr << "ERROR: " << config_file_path << " file doesn't exists\n";
-        return 1;
-    }
-
-    YAML::Node doc = YAML::LoadFile(config_file.string());
-
-    for (YAML::const_iterator it = doc["diagrams"].begin();
-         it != doc["diagrams"].end(); ++it) {
-        if (it->first.as<std::string>() == name) {
-            std::cerr << "ERROR: " << config_file_path
-                      << " file already contains '" << name << "' diagram";
-            return 1;
-        }
-    }
-
-    if (type == clanguml::common::model::diagram_t::kClass) {
-        doc["diagrams"][name]["type"] = "class";
-        doc["diagrams"][name]["glob"] = std::vector<std::string>{{"src/*.cpp"}};
-        doc["diagrams"][name]["using_namespace"] =
-            std::vector<std::string>{{"myproject"}};
-        doc["diagrams"][name]["include"]["namespaces"] =
-            std::vector<std::string>{{"myproject"}};
-        doc["diagrams"][name]["exclude"]["namespaces"] =
-            std::vector<std::string>{{"myproject::detail"}};
-    }
-    else if (type == clanguml::common::model::diagram_t::kSequence) {
-        doc["diagrams"][name]["type"] = "sequence";
-        doc["diagrams"][name]["glob"] = std::vector<std::string>{{"src/*.cpp"}};
-        doc["diagrams"][name]["combine_free_functions_into_file_participants"] =
-            true;
-        doc["diagrams"][name]["using_namespace"] =
-            std::vector<std::string>{{"myproject"}};
-        doc["diagrams"][name]["include"]["paths"] =
-            std::vector<std::string>{{"src"}};
-        doc["diagrams"][name]["exclude"]["namespaces"] =
-            std::vector<std::string>{{"myproject::detail"}};
-        doc["diagrams"][name]["start_from"] =
-            std::vector<std::map<std::string, std::string>>{
-                {{"function", "main(int,const char **)"}}};
-    }
-    else if (type == clanguml::common::model::diagram_t::kPackage) {
-        doc["diagrams"][name]["type"] = "package";
-        doc["diagrams"][name]["glob"] = std::vector<std::string>{{"src/*.cpp"}};
-        doc["diagrams"][name]["using_namespace"] =
-            std::vector<std::string>{{"myproject"}};
-        doc["diagrams"][name]["include"]["namespaces"] =
-            std::vector<std::string>{{"myproject"}};
-        doc["diagrams"][name]["exclude"]["namespaces"] =
-            std::vector<std::string>{{"myproject::detail"}};
-    }
-    else if (type == clanguml::common::model::diagram_t::kInclude) {
-        doc["diagrams"][name]["type"] = "include";
-        doc["diagrams"][name]["glob"] = std::vector<std::string>{{"src/*.cpp"}};
-        doc["diagrams"][name]["relative_to"] = ".";
-        doc["diagrams"][name]["include"]["paths"] =
-            std::vector<std::string>{{"src"}};
-    }
-
-    YAML::Emitter out;
-    out.SetIndent(2);
-
-    out << doc;
-    out << YAML::Newline;
-
-    std::ofstream ofs(config_file);
-    ofs << out.c_str();
-    ofs.close();
-
-    return 0;
-}
-
-void print_config(const clanguml::config::config &cfg)
-{
-    YAML::Emitter out;
-    out.SetIndent(2);
-
-    out << cfg;
-    out << YAML::Newline;
-
-    std::cout << out.c_str();
 }